fix: TrackTime and SoftDelete
