+++ www/api/shared/infraestructure/utils/__init__.py
@@ -1,4 +1,4 @@
 from .api_response import api_response
 from .db_connection import get_db_engine
-from .sqlmodel_to_dict import as_dict, as_json_dumps
+from .sqlmodel_utils import as_dict, as_json_dumps, TrackTimeMixin, SoftDeleteMixin
 from .validate_dict import validate_dict, VKOptions
